backend/requirements.txt
# Core Django and API dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication and JWT
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3

# Database and ORM
psycopg2-binary==2.9.9
django-environ==0.11.2

# Media and file handling
Pillow==10.1.0
django-imagekit==5.0.0

# Utilities and development
python-decouple==3.8
django-debug-toolbar==4.2.0
factory-boy==3.3.0

# Production and deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Rich text and content
django-ckeditor==6.7.0
django-markdownx==4.0.7

# Analytics and monitoring
django-ratelimit==4.1.0

# Testing
pytest-django==4.7.0
pytest==7.4.3
pytest-cov==4.1.0 